Add wheel hub ring and split enemies DB into category modules
